Latest code update with tested all api

+++ parcel_pro/models/sale_order.py
@@ -111,43 +111,47 @@ class SaleOrder(models.Model):
 
     @api.multi
     def action_confirm(self):
-        if not self.order_line:
-            raise ValidationError(_('There should be at least one product in line.'))
-        res = super(SaleOrder, self).action_confirm()
+        p_excep = self.env['parcel.pro.exceptions']
         for order in self:
             if order.carrier_id.parcel_pro == True:
-                self.env['parcel.configuration'].post_quotation(order)
+                print("In actio confirm....",order)
+                if not self.order_line:
+                    p_excep.create({'name': order.name, 'api_type': 'post_quotation',
+                                    'message': "There should be at least one product in line."})
+                    return False
+                    # raise ValidationError(_('There should be at least one product in line.'))
+                post_quotation = self.env['parcel.configuration'].post_quotation(order)
+                print("Post Quote.................",post_quotation)
+                if not post_quotation:
+                    return False
+                res = super(SaleOrder, self).action_confirm()
                 for pick in order.picking_ids:
                     print("pick........",pick)
-                    pick.write({'QuoteId':order.QuoteId,'ParcelPro':order.ParcelPro,'IsHighValueShipment':order.IsHighValueShipment,
+                    pick.write({'QuoteId':order.QuoteId,'ParcelPro':True,'IsHighValueShipment':order.IsHighValueShipment,
                     'IsHighValueShipmentPosted':order.IsHighValueShipmentPosted,'IsCod':order.IsCod,'CarrierCode':order.CarrierCode,
                       'ServiceCode':order.ServiceCode.id,'PackageCode':order.PackageCode.id,'ShipDate':order.ShipDate,
-                                'Height' :order.Height, 'Width' :order.Width,'Length' :order.Length, 'Weight' :order.Weight,
-                    })
-        return res
+                                'Height' :order.Height, 'Width' :order.Width,'Length' :order.Length, 'Weight' :order.Weight})
+                return res
+            else:
+                return  super(SaleOrder, self).action_confirm()
 
     @api.model
     def process_create_quotation(self, ids=None):
-        filters = [('QuoteIdCreated', '=', False),('state', '=', 'draft'),('carrier_id.parcel_pro', '=', True)]
+        filters = [('state', '=', 'draft'),('QuoteIdCreated', '=', False),('carrier_id.parcel_pro', '=', True)]
         order_rec = self.search(filters)
-        print("==order_rec=====",order_rec)
+        print("=========== order_rec ============================",order_rec)
         res = None
-        try:
-            for order in order_rec:
-                print("@@ order @@",order)
-                p_ids = self.env['parcel.pro.exceptions'].search([])
+        if order_rec:
+            try:
+                p_ids = self.env['parcel.pro.exceptions'].search([('api_type', '=', 'post_quotation')])
                 p_ids.unlink()
-                order.action_confirm()
-                # self.env['parcel.configuration'].post_quotation(order)
-        except Exception as e:
-            _logger.exception("Failed processing %s "%e)
+                for order in order_rec:
+                    print("@@ process_create_quotation order @@@@@@@@@@@@@@@@@@@.............",order)
+                    order.action_confirm()
+                    # self.env['parcel.configuration'].post_quotation(order)
+            except Exception as e:
+                _logger.exception("Failed processing %s "%e)
         return res
-
-
-# class SaleOrderLine(models.Model):
-#     _inherit = 'sale.order.line'
-#
-#     InsuredValue = fields.Float('InsuredValue')
 
 class Estimator(models.Model):
     _name = 'estimator'
